Markdown-driven approach
All Standard Record documentation is in an (almost) standard Markdown.
A `CAF.dox` page collects all of them into a single page, as a "topic"
and also as a "related page".
The "(almost)" is a single Doxygen command at the top of the file
which bestows structure and a material for table of contents to `CAF.dox` page.